requirements.txt
python-pptx>=0.6.23
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
requests>=2.31.0
